country overview: distinguish icons by colour
This is ugly -- I don't like having icons on the same page, some of
which are hover-activated, and some of which are click-activated.
Changing that is a larger piece of work, however, so for the minute just
distinguish between the two kinds of icons.

Hover icons remain grey, while clickable icons fall back to being the
same color as link anchors.

Fixes #226.