Update local_census_instances.md
I just moved the regional list to the top, above the cities list.

This puts two excellent and well used examples at the top of the list. There are several city level instances that have been booted, but have no data in them and thus looks bad if someone is hunting for an example.