Move 'show more' button just below hidden text
Having the button at the bottom makes it more obviously clickable since it's right below what you're reading. You've reached this far with your eyes, now you want to read more, and the button is just there.

When it's above the text, I always land up clicking the button for the item *below*, because that's nearest to where my eye is when I reach the part of the text that is hidden.