Switch eslint preset to google.
From airbnb to google. We used google as the base preset for the previous
jscs linter. Many of OKI's JS projects use airbnb, but the census codebase
will need a large style refactor to switch to airbnb, so we'll stick
with google for now, and make project-specific exceptions as and when it
makes sense, to better align with other projects.