[#753] QuestionForm.canAssignProperties method
The conditional check to see whether a dependency should update its
visible properties based on the value of the provider's current state.

Conditions in the schema can be a simple value check; is the dependency
value the same as the provider value. Or a `isNotEmpty` check; where the
provider value must not be empty: string, array or object.