[#753] Pass a context object to the QuestionForm.
We want to be able to pass context data to the QuestionForm. In this
case, it will be data from the dataset we're interested in. First of
all, we'll want the dataset characteristics in the context object, so we
can render the (yet to be committed) multiple choice question type.