[#753] Add QuestionFieldMultipleChoice component
Adds a multiple choice react component for the entry form. Options can
be defined in one of two ways:

A) An array on the `question.config.options` key
B) An array in the context object, the key for which is defined in
`question.config.optionsContextKey` (this is used to get a list of
characteristics from the dataset context object for the
'characteristics' question).