[#784] Submit entry form with javascript.
We have two types of input field in the entry form: 1) Fields controlled
by React, 2) Fields not controlled by React.

We want to gather together the data from the React-controlled fields in
a special way; keeping values and comment values together. So we create
a form and append hidden fields with name/values contructed in the way
we want to send them to the server from
QuestionForm.state.questionState. Other fields, not controlled by React
are also added as hidden fields, mapping their name/values directly. The
newly created form is appended to the doc and submitted.