[#829] Entry form validation
Each `QuestionForm` has a `validate()` method which queries each child
`QuestionField`. The shared `BaseQuestionField` component performs the
guts of the validation, based on a list of `validationRules` passed as a
prop to the component. The `required` rule is an exception as it is
derived from the component's `props.visibleProperties`. Currently only
`required` is supported.

There is basic error message output for each field in the form of a
`helper.QuestionError` component.

Field validation occurs when the value changes or when the form is
submitted. The form won't submit if validation doesn't pass.