[#844] Load dataset.name in preference to .title
The database uses `name` as the field to hold the title value. So the
`title` column in the CMS was mapped to the db `name` field. This meant
`Title@LC` translation columns in the CMS didn't work as expected. This
commit updates the MigrationNotes doc to warn that `Name` should be used
in the CMS instead of `Title`. Title is still loadable, but won't be
translatable.