[#853] Get Survey data directly from codebase
Instead of using metalsmith-request to get data from an API call, use
the census.models.utils.getData() method within a metalsmith plugin to
request it (and modify where necessary) the data directly. This ensures
the data is used in the same way as in the Survey, and gives access to
the full range of model instance methods.