[#854] Push site to AWS S3 bucket
If the deploy flag is added to the generate command, (`--deploy` or `-d`),
write files to the an S3 bucket. The bucket is created as needed with
the name of the index site, e.g. global-test will result in a bucket
name of global-test-index.okfn.org. The bucket is set up with a static
site configuration and global read permissions.