[#856] Add an optional subdomain suffix from env
The generate index script looks for an `INDEX_DOMAIN_SUFFIX` process.env
and adds it to the subdomain used for the S3 bucket name. e.g. if the
`INDEX_DOMAIN_SUFFIX` is `staging` the bucket name for the `global-test`
site would be `global-test-index-staging.okfn.org`.

This can be used to prevent static sites being overwritten on AWS by
generation of sites on staging or development.

A new `generate_index_dev` command has been added to package.json that
adds the `INDEX_DOMAIN_SUFFIX` key with the value `dev`. The staging
server is expected to have a corresponding value for the key set up in
its env.